allocator: mv caching_allocator into separate file
